+++ lib/apis/sign_api/i_sign_engine_wallet.dart
@@ -12,7 +12,6 @@ import 'package:walletconnect_flutter_v2/apis/sign_api/models/sign_client_models
 abstract class ISignEngineWallet extends ISignEngineCommon {
   abstract final Event<SessionProposalEvent> onSessionProposal;
   abstract final Event<SessionRequestEvent> onSessionRequest;
-  abstract final Event<SessionProposalEvent> onProposalExpire;
 
   Future<PairingInfo> pair({
     required Uri uri,
@@ -50,6 +49,5 @@ abstract class ISignEngineWallet extends ISignEngineCommon {
   SessionData? find({
     required Map<String, RequiredNamespace> requiredNamespaces,
   });
-  Map<String, ProposalData> getPendingSessionProposals();
   Map<String, SessionRequest> getPendingSessionRequests();
 }